cleanup: remove unused functions, move handlers
